+++ frontend/src/app/(routes)/staking/components/StakingPage.tsx
@@ -1,6 +1,6 @@
 'use client';
 
-import React, { useEffect, useState } from 'react';
+import React, { useEffect } from 'react';
 import { useAppDispatch, useAppSelector } from '@/custom-hooks/StateHooks';
 import { RootState } from '@/store/store';
 import {
@@ -8,7 +8,6 @@ import {
   getDelegations,
   getParams,
   getUnbonding,
-  txDelegate,
 } from '@/store/features/staking/stakeSlice';
 
 import ChainDelegations from './ChainDelegations';
@@ -16,12 +15,8 @@ import StakingSidebar from './StakingSidebar';
 import ChainUnbondings from './ChainUnbondings';
 import { getDelegatorTotalRewards } from '@/store/features/distribution/distributionSlice';
 import { Validator } from '@/types/staking';
-import DialogDelegate from './DialogDelegate';
-import DialogUndelegate from './DialogUndelegate';
-import DialogRedelegate from './DialogRedelegate';
 import { useRouter } from 'next/navigation';
 import { getBalances } from '@/store/features/bank/bankSlice';
-import { parseBalance } from '@/utils/denom';
 
 const StakingPage = ({
   chainName,
@@ -55,15 +50,6 @@ const StakingPage = ({
   const rewards = useAppSelector(
     (state: RootState) =>
       state.distribution.chains?.[chainID]?.delegatorRewards.list
-  );
-  const stakingParams = useAppSelector(
-    (state: RootState) => state.staking.chains[chainID]?.params
-  );
-  const balance = useAppSelector(
-    (state: RootState) => state.bank.balances[chainID]
-  );
-  const txStatus = useAppSelector(
-    (state: RootState) => state.staking.chains[chainID]?.tx
   );
 
   const allChainInfo = networks[chainID];
@@ -110,92 +96,10 @@ const StakingPage = ({
     dispatch(getParams({ baseURL, chainID }));
   }, [chainID]);
 
-  const [availableBalance, setAvailableBalance] = useState(0);
-  useEffect(() => {
-    if (chainInfo.config.currencies.length > 0) {
-      setAvailableBalance(
-        parseBalance(
-          balance?.list?.length ? balance.list : [],
-          currency.coinDecimals,
-          currency.coinMinimalDenom
-        )
-      );
-    }
-  }, [balance]);
-
-  const [delegateOpen, setDelegateOpen] = useState(false);
-  const [undelegateOpen, setUndelegateOpen] = useState(false);
-  const [redelegateOpen, setRedelegateOpen] = useState(false);
-  const [selectedValidator, setSelectedValidator] = useState<Validator>();
-
-  const handleDialogClose = () => {
-    router.push('/staking/passage');
-    setDelegateOpen(false);
-    setUndelegateOpen(false);
-    setRedelegateOpen(false);
-  };
-
   const onMenuAction = (type: string, validator: Validator) => {
     const valAddress = validator.operator_address;
-
-    setSelectedValidator(validator);
-
-    switch (type) {
-      case 'delegate':
-        setDelegateOpen(true);
-        break;
-      case 'undelegate':
-        setUndelegateOpen(true);
-        break;
-      case 'redelegate':
-        setRedelegateOpen(true);
-        break;
-      default:
-        console.log('unsupported type');
-    }
     router.push(`?validator_address=${valAddress}&action=${type}`);
   };
-
-  const onDelegateTx = (data: { validator: string; amount: number }) => {
-    dispatch(
-      txDelegate({
-        basicChainInfo: {
-          baseURL: baseURL,
-          chainID: chainID,
-          aminoConfig: chainInfo.aminoConfig,
-          rest: chainInfo.config.rest,
-          rpc: chainInfo.config.rpc,
-        },
-        delegator: address,
-        validator: data.validator,
-        amount: data.amount * 10 ** currency.coinDecimals,
-        denom: currency.coinMinimalDenom,
-        prefix: chainInfo.config.bech32Config.bech32PrefixAccAddr,
-        feeAmount:
-          (chainInfo?.config?.feeCurrencies?.[0]?.gasPriceStep?.average || 0) *
-          10 ** currency?.coinDecimals,
-        feegranter: '',
-      })
-    );
-  };
-
-  useEffect(() => {
-    if (validatorAddress.length && action.length && validators?.active) {
-      const validatorInfo =
-        validators.active[validatorAddress] ||
-        validators.inactive[validatorAddress] ||
-        {};
-      const validatorExist = Object.keys(validatorInfo).length ? true : false;
-      setSelectedValidator(validatorInfo);
-      if (action === 'delegate' && validatorExist) {
-        setDelegateOpen(true);
-      } else if (action === 'undelegate' && validatorExist) {
-        setUndelegateOpen(true);
-      } else if (action === 'redelegate' && validatorExist) {
-        setRedelegateOpen(true);
-      }
-    }
-  }, [validatorAddress, action, validators]);
 
   return (
     <div className="flex justify-between">
@@ -209,7 +113,9 @@ const StakingPage = ({
             validators={validators}
             currency={currency}
             rewards={rewards}
-            onMenuAction={onMenuAction}
+            validatorAddress={validatorAddress}
+            action={action}
+            chainSpecific={true}
           />
         </div>
 
@@ -232,18 +138,6 @@ const StakingPage = ({
         currency={currency}
         onMenuAction={onMenuAction}
       />
-      <DialogDelegate
-        onClose={handleDialogClose}
-        open={delegateOpen}
-        validator={selectedValidator}
-        stakingParams={stakingParams}
-        availableBalance={availableBalance}
-        loading={txStatus?.status}
-        displayDenom={currency.coinDenom}
-        onDelegate={onDelegateTx}
-      />
-      <DialogUndelegate onClose={handleDialogClose} open={undelegateOpen} />
-      <DialogRedelegate onClose={handleDialogClose} open={redelegateOpen} />
     </div>
   );
 };
